fix(data): convert type of id from Long to Int
Let's hope user doesn't have more than 2**31 alarms

+++ app/src/main/java/io/github/kirasok/alarmix/data/source/AlarmDao.kt
@@ -17,7 +17,7 @@ interface AlarmDao {
   suspend fun getAlarmById(id: Int): Alarm
 
   @Insert(onConflict = OnConflictStrategy.REPLACE)
-  suspend fun insertAlarm(alarm: Alarm) : Int
+  suspend fun insertAlarm(alarm: Alarm) : Long
 
   @Delete
   suspend fun deleteAlarm(alarm: Alarm)
